move rdkafka log to debug
Premise is that, we use info level for captuign events in the service that are significant to the application's business purpose, whereas debug level can be used for logging messages that aid developers to identify issues during a debugging sessions.

logs from rdkafka are too technical to be put in info, as info will contain business logic observed on staging, so moving this to debug should help to reduce the noise.